requirements.txt

PyQt5==5.15.10
pywin32==306
pandas==1.3.5
numpy==1.21.6
openpyxl==3.1.2
SQLAlchemy==1.4.52
python-dotenv==1.0.1
PyYAML==6.0.1
python-telegram-bot==13.15
